agent/requirements.txt
langgraph>=0.2.0
langchain-openai>=0.1.0
langchain-core>=0.2.0
supabase>=2.4.0
psycopg2-binary>=2.9.9
streamlit>=1.35.0
plotly>=5.22.0
pandas>=2.2.0
python-dotenv>=1.0.0
tabulate>=0.9.0
